requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
dj-database-url==0.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djoser==2.0.5